+++ packages/transformer/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @ts-linq/transformer
 
+## 2.1.19
+
+### Patch Changes
+
+- Updated dependencies
+  - @ts-linq/types@4.2.0
+  - @ts-linq/ast@3.2.2
+
 ## 2.1.18
 
 ### Patch Changes